extract game_state
